backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pyjwt[cryptography]
asyncpg
aiosqlite
pydantic-settings
python-dotenv
passlib[bcrypt]
python-multipart
openai
cohere
email-validator
